email-outreach-automation/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

python-dotenv==1.0.1
pydantic==2.7.1

pandas==2.2.2

google-generativeai==0.7.2

resend==0.8.0

celery==5.3.6
redis==5.0.4

requests==2.32.3
beautifulsoup4==4.12.3

python-multipart==0.0.9